+++ tests/stdlib/tstrutils.nim
@@ -3,7 +3,7 @@ discard """
 """
 
 import std/strutils
-from stdtest/testutils import disableVm
+from stdtest/testutils import disableVm, whenVMorJs
 # xxx each instance of `disableVm` and `when not defined js:` should eventually be fixed
 
 template rejectParse(e) =
@@ -13,6 +13,36 @@ template rejectParse(e) =
   except ValueError: discard
 
 template main() =
+  # Tests for the sets
+  
+  whenVMorJs: discard
+  do:
+    block: # Whitespace
+      proc isspace(c: cint): cint {.importc, header: "<ctype.h>".}
+      for i in char.low .. char.high:
+        doAssert (isspace(i.cint) > 0) == (i in Whitespace)
+
+    block: # ControlChars
+      proc iscntrl(c: cint): cint {.importc, header: "<ctype.h>".}
+      for i in char.low .. char.high:
+        doAssert (iscntrl(i.cint) > 0) == (i in ControlChars)
+    
+    block: # GraphicChars
+      proc isgraph(c: cint): cint {.importc, header: "<ctype.h>".}
+      for i in char.low .. char.high:
+        doAssert (isgraph(i.cint) > 0) == (i in GraphicChars)
+    
+    block: # PrintableChars
+      proc isprint(c: cint): cint {.importc, header: "<ctype.h>".}
+      for i in char.low .. char.high:
+        doAssert (isprint(i.cint) > 0) == (i in PrintableChars)
+
+    block: # Punctuation
+      proc ispunct(c: cint): cint {.importc, header: "<ctype.h>".}
+      for i in char.low .. char.high:
+        doAssert (ispunct(i.cint) > 0) == (i in Punctuation)
+
+
   block: # strip
     doAssert strip("  ha  ") == "ha"
     doAssert strip("  foofoofoo  ") == "foofoofoo"